feat: date(util) 함수 작성 #14
- 당일(0시 0분)과 일정 날짜(매개변수)가 지난 날을 반환하는  함수 작성
- order repository 로직 getRangeByToday로 추상화
